refactor gazetteer
